backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
PyJWT==2.8.0

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP Client
httpx==0.26.0

# Email
python-dotenv==1.0.0
emails==0.6

# Utilities
python-dateutil==2.8.2
